requirements.txt
notion-client>=2.2.1
openai>=1.35.0
python-dotenv>=1.0.1
aiohttp>=3.9.0